audience, campaign, subscriber API, CSV upload

+++ controllers/audienceController.js
@@ -1,4 +1,6 @@
 const connection = require("../models/db");
+const fs = require("fs");
+const csv = require("csv-parser");
 
 // Get audience API
 const getAudience = (req, res) => {
@@ -22,10 +24,18 @@ const getAudienceById = (req, res) => {
 // Delete an audience API
 const deleteAudience = (req, res) => {
   connection.query(
-    `delete from audience where id = ${req.params.id}`,
+    `select * from audience where user_id=${req.user.data.userId}`,
     (err, resp) => {
-      if (err) return res.send(err);
-      res.send("audience successfully deleted at ID " + req.params.id);
+      let audience = resp.find((x) => x.id == req.params.id);
+      if (audience == undefined)
+        return res.status(401).send("Cannot delete audience you didn't create");
+      connection.query(
+        `delete from audience where id = ${req.params.id}`,
+        (err, resp) => {
+          if (err) return res.send(err);
+          res.send("audience successfully deleted at ID " + req.params.id);
+        }
+      );
     }
   );
 };
@@ -34,7 +44,6 @@ const deleteAudience = (req, res) => {
 const createAudience = (req, res) => {
   if (
     !req.body.user_id ||
-    !req.body.subscriberGroup ||
     !req.body.firstName ||
     !req.body.lastName ||
     !req.body.email ||
@@ -49,8 +58,8 @@ const createAudience = (req, res) => {
   // INSERT into database
   connection.query(
     `insert into audience (user_id,subscriberGroup,firstName,lastName,tel,email,country,state,city,birthday) values 
-              ('${req.body.user_id}',
-              '${req.body.subscriberGroup}',
+              ('${req.user.data.userId}',
+              '${null}',
               '${req.body.firstName}',
               '${req.body.lastName}',
               '${req.body.tel}',
@@ -70,12 +79,58 @@ const createAudience = (req, res) => {
 //rest api to update record into mysql database
 const editAudience = (req, res) => {
   connection.query(
-    `update audience set firstName = '${req.body.firstName}', lastName = '${req.body.lastName}', tel='${req.body.tel}' where id=${req.params.id}`,
-    (err, response) => {
-      if (err) throw err;
-      res.send("audience edited Successfully");
+    `select * from audience where user_id=${req.user.data.userId}`,
+    (err, resp) => {
+      let audience = resp.find((x) => x.id == req.params.id);
+      if (audience == undefined)
+        return res.status(401).send("Cannot edit audience you didn't create");
+      connection.query(
+        `update audience set firstName = '${req.body.firstName}', lastName = '${req.body.lastName}', tel='${req.body.tel}', email='${req.body.email}',
+         country='${req.body.country}', state='${req.body.state}', city='${req.body.city}', birthday='${req.body.birthday}' where id=${req.params.id}`,
+        (err, response) => {
+          if (err) throw err;
+          res.send("audience edited Successfully");
+        }
+      );
     }
   );
+};
+
+// Upload CSV file
+const importContacts = (req, res) => {
+  if (req.file == undefined)
+    return res.status(400).send("Please upload a CSV file!");
+
+  const results = [];
+  let path = "uploads/docs/" + req.file.filename;
+
+  fs.createReadStream(path)
+    .pipe(csv())
+    .on("data", (data) => results.push(data))
+    .on("end", () => {
+      let values = "values";
+
+      results.map((contact) => {
+        if (contact.birthday == undefined) contact.birthday = null;
+        if (contact.country == undefined) contact.country = contact.county;
+
+        values += `(${req.user.data.userId}, '${null}', '${
+          contact.first_name
+        }', '${contact.last_name}', '${contact.phone}', '${contact.email}', '${
+          contact.country
+        }', '${contact.state}', '${contact.city}', '${contact.birthday}'),\n`;
+      });
+
+      let finalValue = values.slice(0, -2);
+
+      connection.query(
+        `insert into audience (user_id,subscriberGroup,firstName,lastName,tel,email,country,state,city,birthday) ${finalValue}`,
+        (dbErr, resp) => {
+          if (dbErr) return res.status(400).send(dbErr);
+          res.send("Contacts Imported Successfully");
+        }
+      );
+    });
 };
 
 module.exports = {
@@ -84,4 +139,5 @@ module.exports = {
   deleteAudience,
   createAudience,
   editAudience,
+  importContacts,
 };
