+++ controllers/subscriberGroup.js
@@ -22,10 +22,22 @@ const getSubscriberGroupById = (req, res) => {
 // Delete an subscriber_group API
 const deleteSubscriberGroup = (req, res) => {
   connection.query(
-    `delete from subscriber_group where id = ${req.params.id}`,
+    `select * from subscriber_group where userId=${req.user.data.userId}`,
     (err, resp) => {
-      if (err) return res.send(err);
-      res.send("subscriber_group successfully deleted at ID " + req.params.id);
+      let subscriberGroup = resp.find((x) => x.id == req.params.id);
+      if (subscriberGroup == undefined)
+        return res
+          .status(401)
+          .send("Cannot delete subscriber group you didn't create");
+      connection.query(
+        `delete from subscriber_group where id = ${req.params.id}`,
+        (err, resp) => {
+          if (err) return res.send(err);
+          res.send(
+            "subscriber_group successfully deleted at ID " + req.params.id
+          );
+        }
+      );
     }
   );
 };
@@ -37,12 +49,12 @@ const createSubscriberGroup = (req, res) => {
 
   // INSERT into database
   connection.query(
-    `insert into subscriber_group (name,user_id) values 
+    `insert into subscriber_group (name,userId) values 
                 ('${req.body.name}',
                 '${req.user.data.userId}')`,
     (error, resp) => {
       if (error) return res.send(error.sqlMessage);
-      res.send("subscriber_group successfully created.");
+      res.send("subscriber group successfully created.");
       res.end();
     }
   );
@@ -51,10 +63,20 @@ const createSubscriberGroup = (req, res) => {
 //rest api to update record into mysql database
 const editSubscriberGroup = (req, res) => {
   connection.query(
-    `update subscriber_group set user_id = '${req.user.data.userId}', name = '${req.body.name}' where id=${req.params.id}`,
-    (err, response) => {
-      if (err) throw err;
-      res.send("subscriber_group edited Successfully");
+    `select * from subscriber_group where userId=${req.user.data.userId}`,
+    (err, resp) => {
+      let subscriberGroup = resp.find((x) => x.id == req.params.id);
+      if (subscriberGroup == undefined)
+        return res
+          .status(401)
+          .send("Cannot edit subscriber group you didn't create");
+      connection.query(
+        `update subscriber_group set name = '${req.body.name}' where id=${req.params.id}`,
+        (err, response) => {
+          if (err) throw err;
+          res.send("subscriber group edited Successfully");
+        }
+      );
     }
   );
 };
